requirements.txt
aiogram==2.25.2
pyyaml==6.0.0